feat: Implement of EventTarget
